Фикс багов
