+++ src/main/java/faang/school/projectservice/controller/recommendation/ProjectController.java
@@ -1,22 +1,21 @@
 package faang.school.projectservice.controller.recommendation;
 
-import faang.school.projectservice.config.context.UserContext;
-import faang.school.projectservice.dto.project.ProjectFilterDto;
+import faang.school.projectservice.config.feign.UserContext;
 import faang.school.projectservice.dto.project.ProjectCreateDto;
+import faang.school.projectservice.dto.project.ProjectFilterDto;
 import faang.school.projectservice.dto.project.ProjectReadDto;
 import faang.school.projectservice.dto.project.ProjectUpdateDto;
 import faang.school.projectservice.service.ProjectManagementService;
+import faang.school.projectservice.service.ProjectService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 
 import java.util.List;
 
@@ -27,6 +26,7 @@ public class ProjectController {
 
     private final ProjectManagementService projectManagementService;
     private final UserContext userContext;
+    private final ProjectService projectService;
 
     @PostMapping
     public ProjectReadDto createProject(
@@ -51,6 +51,16 @@ public class ProjectController {
     public ProjectReadDto getProjectById(
             @PathVariable long projectId ){
         return projectManagementService.getProjectById(projectId, userContext.getUserId());
+    }
+
+    @PostMapping("/{projectId}/presentation")
+    public ProjectReadDto generatePdf(@PathVariable long projectId) {
+        return projectService.generateProjectPresentation(projectId);
+    }
+
+    @GetMapping("/{projectId}/presentation")
+    public String getPresentationFilePath(@PathVariable long projectId) {
+        return projectService.getPresentationFileKey(projectId);
     }
 
 }
