+++ src/main/resources/db/changelog/changeset/project_V016_meet.sql
@@ -8,12 +8,14 @@ CREATE TABLE IF NOT EXISTS meet (
     starts_at TIMESTAMP NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    CONSTRAINT fk_meet_project FOREIGN KEY (project_id) REFERENCES project (id)
+    CONSTRAINT fk_meet_project FOREIGN KEY (project_id) REFERENCES project (id),
+    CONSTRAINT fk_meet_user FOREIGN KEY (creator_id) REFERENCES users (id)
 );
 
 CREATE TABLE IF NOT EXISTS meet_participant (
     id BIGSERIAL PRIMARY KEY,
     meet_id BIGINT NOT NULL,
     user_id BIGINT NOT NULL,
-    CONSTRAINT fk_participant_meet FOREIGN KEY (meet_id) REFERENCES meet (id)
+    CONSTRAINT fk_participant_meet FOREIGN KEY (meet_id) REFERENCES meet (id),
+    CONSTRAINT fk_participant_user FOREIGN KEY (user_id) REFERENCES users (id)
 );
